VM: Update API with new changes
gas changed to long and transactionResult changed

+++ src/main/java/org/semux/vm/client/SemuxTransaction.java
@@ -53,8 +53,8 @@ public class SemuxTransaction implements Transaction {
     }
 
     @Override
-    public BigInteger getGas() {
-        return BigInteger.valueOf(transaction.getGas());
+    public long getGas() {
+        return transaction.getGas();
     }
 
     @Override
